refactor getCentreAtom to utils
