+++ baby-gru/src/components/card/MoorhenMoleculeCard.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState, useRef, useReducer, useCallback, useImperativeHandle, forwardRef } from 'react';
 import { Card, Row, Col, Stack, Button, Spinner } from "react-bootstrap";
 import { Accordion, AccordionDetails, AccordionSummary } from '@mui/material';
-import { convertRemToPx, convertViewtoPx, representationLabelMapping } from '../../utils/MoorhenUtils';
+import { convertRemToPx, convertViewtoPx, getCentreAtom, representationLabelMapping } from '../../utils/MoorhenUtils';
 import { isDarkBackground } from '../../WebGLgComponents/mgWebGL'
 import { MoorhenSequenceList } from "../list/MoorhenSequenceList";
 import { MoorhenMoleculeCardButtonBar } from "../button-bar/MoorhenMoleculeCardButtonBar"
@@ -19,7 +19,6 @@ import { addMolecule } from '../../store/moleculesSlice';
 import { showMolecule } from '../../store/moleculeRepresentationsSlice';
 import { triggerUpdate } from '../../store/moleculeMapUpdateSlice';
 import { MoorhenCarbohydrateList } from "../list/MoorhenCarbohydrateList";
-import { libcootApi } from '../../types/libcoot';
 
 const allRepresentations = [ 'CBs', 'CAs', 'CRs', 'ligands', 'gaussian', 'MolecularSurface', 'DishyBases', 'VdwSpheres', 'rama', 'rotamer', 'CDs', 'allHBonds','glycoBlocks', 'restraints' ]
 
@@ -149,14 +148,8 @@ export const MoorhenMoleculeCard = forwardRef<any, MoorhenMoleculeCardPropsInter
     }
 
     const redrawAdaptativeBonds = async () => {
-        const response = await props.commandCentre.current.cootCommand({
-            returnType: "int_string_pair",
-            command: "get_active_atom",
-            commandArgs: [...props.glRef.current.origin.map(coord => coord * -1), props.molecule.molNo.toString()]
-        }, false) as moorhen.WorkerResponse<libcootApi.PairType<number, string>>
-        const moleculeMolNo: number = response.data.result.result.first
-        const residueCid: string = response.data.result.result.second
-        if (moleculeMolNo === props.molecule.molNo) {
+        const [molecule, residueCid] = await getCentreAtom([props.molecule], props.commandCentre, props.glRef)
+        if (molecule.molNo === props.molecule.molNo) {
             await props.molecule.drawAdaptativeBonds(residueCid, 10)
         }
     }
@@ -164,7 +157,7 @@ export const MoorhenMoleculeCard = forwardRef<any, MoorhenMoleculeCardPropsInter
     const handleOriginUpdate = useCallback(() => {
         isDirty.current = true
         if (!busyRedrawing.current) {
-            redrawSymmetryIfDirty()
+                        redrawSymmetryIfDirty()
         }
 
     }, [props.molecule, props.glRef])
